requirements.txt
pandas>=2.1.0
numpy>=1.26.0
scikit-learn>=1.4.0
mlflow>=2.10.0
pytest>=7.4.3
pyyaml>=6.0.1
joblib>=1.3.2